+++ CHANGELOG.md
@@ -30,11 +30,10 @@ This release requires regeneration of ThriftRW code.
     when instantiating Thrift clients or registering handlers.
 -   **Breaking**: Use of `golang.org/x/net/context` has been dropped in favor
     of the standard library's `context` package.
--   Rename `Handler` interface to `HandleUnary` and separate Outbound
+-   Rename `Handler` interface to `UnaryHandler` and separate `Outbound`
     interface into `Outbound` and `UnaryOutbound`.
 -   Add `OnewayHandler` and `HandlerSpec` to support oneway handlers.
     Transport inbounds can choose which RPC types to accept
-
 
 v0.3.1 (2016-09-31)
 -------------------